*Datatype: - Added default function canRemovable( $contentClassAttribute ) in ezdatatype.php. This   function is used to check whether or not a datatype can be removed from a class. Function   returns true in default. - Added/reimplemented function canRemovable() in datatype ezuser, disallowed deletion of this   datatype if it belongs to the class of login user.
git-svn-id: file:///home/patrick.allaert/svn-git/ezp-repo/ezpublish/trunk@8604 a01eee8c-daba-0310-acae-fa49f3023285

+++ kernel/classes/ezcontentclassattribute.php
@@ -258,10 +258,16 @@ class eZContentClassAttribute extends eZPersistentObject
     {
         $dataType =& $this->dataType();
         $version = $this->Version;
-        $dataType->deleteStoredClassAttribute( $this, $version );
-        eZPersistentObject::remove();
+        if ( $dataType->canRemovable( $this ) )
+        {
+            $dataType->deleteStoredClassAttribute( $this, $version );
+            eZPersistentObject::remove();
+        }
+        else
+        {
+            eZDebug::writeError( 'Datatype [' . $dataType->Name . '] can not be deleted to avoid system crash' );
+        }
     }
-
 
     function &fetch( $id, $asObject = true, $version = EZ_CLASS_VERSION_STATUS_DEFINED, $field_filters = null )
     {
